A2A agent reorganize files - both api and cli can use common functions (#140)
* Instructions to support both run via cli and run via a2a serving

* Rename available_tools.md

* Factory tools

* Move trace_utils to agent

* AGENT_CODE_TEMPLATE_RUN_VIA_CLI in cli generation

* Rename available_tools.md

* Unify instructions into one

* Unify instructions into one

* Common file for schemas

* load_system_instructions instead of variable

* more functionality to agent.utils

* A2A serving works with new imports

* templates.py not used anymore

* Move save function to io_utils.py

* Rename arg to for_cli_agent

* Undo moving save_agent_outputs

* Note about .env in agent README

* Fix agent_factory imports

* Fix imports in test suite

* Failing CI test: requirements validator to use any-agent[all,a2a]

* Fix test patch

cost compute in instructions
