+++ src/agent/__main__.py
@@ -3,25 +3,11 @@ import fire
 from any_agent import AgentConfig, AgentFramework, AnyAgent
 from any_agent.serving import A2AServingConfig
 from any_agent.tools import search_tavily, visit_webpage
-from instructions import INSTRUCTIONS
-from pydantic import BaseModel, Field
-from utils import read_file, search_mcp_servers  # type: ignore[import-not-found]
+from factory_tools import read_file, search_mcp_servers
+from instructions import load_system_instructions
+from schemas import AgentFactoryOutputs
 
 dotenv.load_dotenv()
-
-
-class AgentFactoryOutputs(BaseModel):
-    agent_instructions: str = Field(..., description="The instructions passed to the generated agent.")
-    tools: str = Field(..., description="The python code that defines the tools to be used by the generated agent.")
-    imports: str = Field(..., description="The python code snippet used to import the required tools.")
-    structured_outputs: str = Field(..., description="The Pydantic v2 models used to structure the agent's output.")
-    cli_args: str = Field(..., description="The arguments to be provided to the agent from the command line.")
-    agent_description: str = Field(..., description="The description of the agent and what it does.")
-    prompt_template: str = Field(
-        ..., description="A prompt template that, completed with cli_args, defines the agent's input prompt."
-    )
-    run_instructions: str = Field(..., description="The run instructions in Markdown format")
-    dependencies: str = Field(..., description="The dependencies line by line in Markdown format")
 
 
 def main(
@@ -51,7 +37,7 @@ def main(
         framework,
         AgentConfig(
             model_id=model,
-            instructions=INSTRUCTIONS,
+            instructions=load_system_instructions(for_cli_agent=False),
             description="Agent for generating agentic workflows based on user prompts.",
             tools=[visit_webpage, search_tavily, search_mcp_servers, read_file],
             model_args={"tool_choice": "required"},  # Ensure tool choice is required
