+++ src/agent/instructions.py
@@ -6,6 +6,28 @@ from jinja2 import Template
 
 ANY_AGENT_VERSION = version("any_agent")
 
+TOOLS_REMINDER = """Use appropriate tools in the agent configuration:
+- Select relevant tools from `tools/README.md`.
+- Use the `search_mcp_servers` tool to discover and add MCP servers that provide relevant tools
+    to the configuration.
+
+Always use the simplest and most efficient tools available for the task.
+"""
+
+USER_PROMPT = """Generate Python code for an agentic workflow using the `any-agent` library
+to do the following:
+{0}
+
+{1}
+"""
+
+AMENDMENT_PROMPT = """
+Amend the Python code you generated for the agentic workflow to do the following:
+{0}
+
+If necessary, {1}
+"""
+
 CODE_EXAMPLE_WITH_COMMENTS = """
 # agent.py
 
@@ -13,12 +35,16 @@ CODE_EXAMPLE_WITH_COMMENTS = """
 import os
 
 # ALWAYS used
+import json
+from pathlib import Path
+from any_agent.serving import A2AServingConfig
 from dotenv import load_dotenv
 from any_agent import AgentConfig, AnyAgent
-from any_agent.serving import A2AServingConfig
-from any_agent.config import MCPStdio
 from pydantic import BaseModel, Field
 from fire import Fire
+
+# MCPStdio should be imported ONLY if MCP servers are used in AgentConfig
+from any_agent.config import MCPStdio
 
 # ADD BELOW HERE: tools made available by any-agent or agent-factory
 from any_agent.tools import visit_webpage
@@ -29,35 +55,24 @@ load_dotenv()
 
 # ========= Structured output definition =========
 class StructuredOutput(BaseModel):
-    url: str = Field(
-        ..., description="The URL of the webpage that was translated.")
-    source_language: str = Field(
-        ..., description="The source language detected on the webpage (should be 'English').")
-    extracted_text: str = Field(
-        ..., description="The main text content extracted from the original English webpage.")
-    translated_text: str = Field(
-        ..., description="The English text translated to Italian.")
+    url: str = Field(..., description="The URL of the webpage that was translated.")
+    source_language: str = Field(..., description="The source language detected on the webpage (should be 'English').")
+    extracted_text: str = Field(..., description="The main text content extracted from the original English webpage.")
+    translated_text: str = Field(..., description="The English text translated to Italian.")
 
 
 # ========= System Instructions =========
 INSTRUCTIONS = '''
-You are an assistant that translates the main text content of an English webpage to Italian,
-following this step-by-step workflow:
-1. Receive a webpage URL from the user. Visit the page and extract the primary and most relevant
-   English text content. Focus on body content, main text, and important sections. Exclude
-   navigation bars, headings not part of the content, footers, advertisements, and non-informational
-   elements. Make sure the extracted text is concise but comprehensive and represents the actual
-   page content.
-2. Identify and confirm that the detected source language is English. If the page is not in English,
-   halt and output the detected language and a clear message in 'translated_text'.
+You are an assistant that translates the main text content of an English webpage to Italian, following this step-by-step workflow:
+1. Receive a webpage URL from the user. Visit the page and extract the primary and most relevant English text content. Focus on body content, main text, and important sections. Exclude navigation bars, headings not part of the content, footers, advertisements, and non-informational elements. Make sure the extracted text is concise but comprehensive and represents the actual page content.
+2. Identify and confirm that the detected source language is English. If the page is not in English, halt and output the detected language and a clear message in 'translated_text'.
 3. Use the translation tool to translate the extracted English text into fluent Italian.
 4. Your output must be a structured JSON object with these fields:
    - url: the provided webpage URL
    - source_language: the detected primary language (should be English)
    - extracted_text: the main English content you extracted
    - translated_text: your Italian translation of the extracted text
-Limit the output to 1000 tokens if the page is very long. Ensure the translation is accurate and
-clear. Do not make up or hallucinate content.
+Limit the output to 1000 tokens if the page is very long. Ensure the translation is accurate and clear. Do not make up or hallucinate content.
 '''
 
 
@@ -86,9 +101,50 @@ TOOLS = [
     ),
 ]
 
+"""  # noqa: E501
 
+CODE_EXAMPLE_RUN_VIA_CLI = """
+# ========== Running the agent via CLI ===========
+agent = AnyAgent.create(
+    "openai",
+    AgentConfig(
+        model_id="o3",
+        instructions=INSTRUCTIONS,
+        tools=TOOLS,
+        output_type=StructuredOutput,
+    ),
+)
+
+
+def main(url: str):
+    \"\"\"
+    Given a webpage URL, translate its main English content to Italian,
+    and return structured output.
+    \"\"\"
+    input_prompt = f"Translate the main text content from the following English webpage URL to Italian: {url}"
+    try:
+        agent_trace = agent.run(prompt=input_prompt, max_turns=20)
+    except AgentRunError as e:
+        agent_trace = e.trace
+        print(f"Agent execution failed: {{str(e)}}")
+        print("Retrieved partial agent trace...")
+
+    script_dir = Path(__file__).resolve().parent
+    output_path = script_dir / "agent_eval_trace.json"
+    with open(output_path, "w", encoding="utf-8") as f:
+        f.write(agent_trace.model_dump_json(indent=2))
+
+    return agent_trace.final_output
+
+
+if __name__ == "__main__":
+    Fire(main)
+"""  # noqa: E501
+
+CODE_EXAMPLE_RUN_VIA_A2A = """
+# ========== Running the agent via A2AServing ===========
 def main(
-    franework: str = "openai",
+    framework: str = "openai",
     model: str = "o3",
     host: str = "localhost",
     port: int = 8080,
@@ -135,45 +191,35 @@ The final expected output is a dictionary with the following structure:
     "structured_outputs": "The Pydantic v2 models used to structure the agent's final output.",
     "cli_args": "The arguments to be provided to the agent from the command line.",
     "agent_description": "The description of the agent and what it does.",
-    "prompt_template": "A prompt template that, completed with cli args, defines the agent's input prompt.",
+    "prompt_template": "A prompt template that, completed with cli_args, defines the agent's input prompt.",
     "run_instructions": "The instructions for setting up the environment in Markdown format (e.g., A README file).",
     "dependencies": "The list of python dependencies in Markdown format."
 }}
 
 ## Values to assign to dictionary keys
 
-1. `agent_instructions` is a string that will be assigned to the `INSTRUCTIONS` variable in the
-   template (type: str). This string replaces the {{agent_instructions}} placeholder in the agent
-   code template.
-2. `tools` is python code that assigns the `TOOLS` variable with the list of tools required by the
-   generated agent. This code replaces the {{tools}} placeholder in the agent code template.
-3. `imports` is python code containing all the required imports for the selected tools. This code
-   replaces the {{imports}} placeholder in the agent code template.
-4. `structured_outputs` is python code that defines the class `StructuredOutput(BaseModel)`)
-   defining the agent's output schema as a Pydantic v2 model. This code replaces the
-   {{structured_outputs}} placeholder in the agent code template.
-5. `cli_args` are the arguments to be passed to the `run_agent` function. Each of them is specified
-   as argument_name: argument_value. These will replace the {{cli_args}} placeholder in the agent
-   code template.
-6. `agent_description` is a string to be provided as the description of the `run_agent` function.
-7. `prompt_template` is an f-string which is formatted with the values of `cli_args` to build the
-   final input prompt to the generated agent.
+1. `agent_instructions` is a string that will be assigned to the `INSTRUCTIONS` variable in the template (type: str).
+This string replaces the {{agent_instructions}} placeholder in the agent code template.
+2. `tools` is python code that assigns the `TOOLS` variable with the list of tools required by the generated agent. This code replaces the {{tools}} placeholder in the agent code template.
+3. `imports` is python code containing all the required imports for the selected tools. This code replaces the {{imports}} placeholder in the agent code template.
+4. `structured_outputs` is python code that defines the class `StructuredOutput(BaseModel)`) defining the agent's output schema as a Pydantic v2 model.
+This code replaces the {{structured_outputs}} placeholder in the agent code template.
+5. `cli_args` are the arguments to be passed to the `main` function. Each of them is specified as argument_name: argument_value.
+These will replace the {{cli_args}} placeholder in the agent code template.
+6. `agent_description` is a string to be provided as the description of the `main` function.
+7. `prompt_template` is an f-string which is formatted with the values of `cli_args` to build the final input prompt to the generated agent.
 8. `run_instructions` should contain clear and concise setup instructions:
-    - Environment variables: Instruct the user to create a .env file to set environment variables;
-      specify exactly which environment variables are required
-    - Always include the following instructions to install Python package manager uv (the end user
-      decides which command to run based on their OS):
+    - Environment variables: Instruct the user to create a .env file to set environment variables; specify exactly which environment variables are required
+    - Always include the following instructions to install Python package manager uv (the end user decides which command to run based on their OS):
         - for MacOS and Linux users: `curl -LsSf https://astral.sh/uv/install.sh | sh`
         - for Windows users: `powershell -ExecutionPolicy ByPass -c "irm https://astral.sh/uv/install.ps1 | iex"`
-    - Run instructions for agent.py using `uv run` with specification of `requirements.txt` and
-      Python 3.13:
-      `uv run --with-requirements requirements.txt --python 3.13 python agent.py --arg1 "value1"`
-9. dependencies should list all the python libraries (including the ones required by the tools) as
-   dependencies to be installed. It will be used to generate the `requirements.txt` file:
-    - The first line should be `any-agent[all,a2a]=={ANY_AGENT_VERSION}` dependency, since we are using
-      any-agent to run the agent workflow
-    - (Optional) If the `agent_code` uses `uvx` to spin up any MCP server, include "uv" as a
-      dependency in the `requirements.txt` file
+    - Run instructions for agent.py using `uv run` with specification of requirements.txt and Python 3.11
+      `uv run --with-requirements generated_workflows/<folder_name>/requirements.txt --python 3.11 python generated_workflows/<folder_name>/agent.py --arg1 "value1"`
+      where the user is expected to replace <folder_name> with the timestamped folder created in the generated_workflows directory and specify the required arguments
+9. dependencies should list all the python libraries (including the ones required by the tools) as dependencies to be installed. It will be used to generate the requirements.txt file
+    - the first line should be "any-agent[all,a2a]=={ANY_AGENT_VERSION}" dependency, since we are using any-agent to run the agent workflow
+    - only if the `agent_code` uses `uvx` to spin up any MCP server, include "uv" as a dependency in the requirements.txt file
+    - do not provide specific versions for the dependencies except for `any-agent[all,a2a]` (see the above point)
 """  # noqa: E501
 
 AGENT_CODE_TEMPLATE = """
@@ -183,10 +229,11 @@ AGENT_CODE_TEMPLATE = """
 import os
 
 # ALWAYS used
+import json
+from pathlib import Path
+from any_agent.serving import A2AServingConfig
 from dotenv import load_dotenv
 from any_agent import AgentConfig, AnyAgent, AgentRunError
-from any_agent.serving import A2AServingConfig
-from any_agent.config import MCPStdio
 from pydantic import BaseModel, Field
 from fire import Fire
 
@@ -206,7 +253,71 @@ INSTRUCTIONS='''
 # ========== Tools definition ===========
 {tools}
 
+"""  # noqa: E501
 
+AGENT_CODE_TEMPLATE_RUN_VIA_CLI = """
+# ========== Running the agent via CLI ===========
+agent = AnyAgent.create(
+    "openai",
+    AgentConfig(
+        model_id="o3",
+        instructions=INSTRUCTIONS,
+        tools=TOOLS,
+        output_type=StructuredOutput,
+        model_args={{"tool_choice": "required"}},
+    ),
+)
+
+def main({cli_args}):
+    \"\"\"{agent_description}\"\"\"
+    input_prompt = f"{prompt_template}"
+    try:
+        agent_trace = agent.run(prompt=input_prompt, max_turns=20)
+    except AgentRunError as e:
+        agent_trace = e.trace
+        print(f"Agent execution failed: {{str(e)}}")
+        print("Retrieved partial agent trace...")
+
+    # Extract cost information (with error handling)
+    try:
+        cost_info = agent_trace.cost
+        if cost_info.total_cost > 0:
+            cost_msg = (
+                f"input_cost=${{cost_info.input_cost:.6f}} + "
+                f"output_cost=${{cost_info.output_cost:.6f}} = "
+                f"${{cost_info.total_cost:.6f}}"
+            )
+    except Exception as e:
+        class DefaultCost:
+            input_cost = 0.0
+            output_cost = 0.0
+            total_cost = 0.0
+        cost_info = DefaultCost()
+
+    # Create enriched trace data with costs as separate metadata
+    script_dir = Path(__file__).resolve().parent
+    output_path = script_dir / "agent_eval_trace.json"
+
+    # Prepare the trace data with costs
+    trace_data = agent_trace.model_dump()
+    trace_data["execution_costs"] = {{
+        "input_cost": cost_info.input_cost,
+        "output_cost": cost_info.output_cost,
+        "total_cost": cost_info.total_cost
+    }}
+
+    with open(output_path, "w", encoding="utf-8") as f:
+        f.write(json.dumps(trace_data, indent=2))
+
+    return agent_trace.final_output
+
+if __name__ == "__main__":
+    Fire(main)
+
+"""  # noqa: E501
+
+AGENT_CODE_TEMPLATE_RUN_VIA_A2A = """
+# ========== Running the agent via A2AServing ===========
 def main(
     framework: str = "openai",
     model: str = "o3",
@@ -232,28 +343,25 @@ def main(
 if __name__ == "__main__":
     Fire(main)
 
-"""
+"""  # noqa: E501
 
 
-CODE_GENERATION_INSTRUCTIONS = f"""
+CODE_GENERATION_INSTRUCTIONS = """
 # Single Agent Implementation with Multiple Steps
 
 ## Task Overview
 Create a complete implementation of a single agent that executes a multi-step workflow
 using Mozilla's any-agent library. The implementation should:
 
-1. Use the OpenAI framework as the underlying agent provider, except if the user specifies a
-   different framework in their request.
-2. Implement a step-by-step approach where the agent breaks down the user's request into multiple
-   steps, each with an input and output.
-3. To obtain JSON output from the agent, define structured output using Pydantic v2 models via the
-   `output_type` argument.
+1. Use the OpenAI framework as the underlying agent provider
+2. Implement a step-by-step approach where the agent breaks down the user's request into multiple steps, each with an input and output
+3. To obtain JSON output from the agent, define structured output using Pydantic v2 models via the `output_type` argument.
 4. Whenever required, assign tools in the agent configuration.
 
 ## Required Components
 
 #### Model (model_id):
-- Use `o3` as the `model_id`, except if the user specifies a different model in their request.
+- Use `o3` as the `model_id`
 
 #### Instructions (instructions):
 - Decide on the number of steps that you think would be necessary to complete the task
@@ -273,17 +381,26 @@ using Mozilla's any-agent library. The implementation should:
        file in the `tools/` directory that implements the function.
     b. Tools pre-defined in any-agent library: `search_tavily` and `visit_webpage` tools
     c. MCP Servers: To discover a relevant MCP server, first use the `search_mcp_servers` tool,
-       giving it a keyword that describes the task you want to accomplish.
-       Then, read each MCP server's description carefully to verify which one provides the tools you
-       need for the task. Each MCP has a configuration that must be accurately implemented in the
-       agent configuration via MCPStdio(). Always suggest only the minimum subset of tools from the
-       MCP server URL that are necessary for the solving the task at hand. If the agent is required
-       to generate any intermediate files, you may ask it to save them in a path relative to the
-       current working directory (do not give absolute paths).
+       giving it a keyphrase that describes the task you want to accomplish.
+       Then, read each MCP server's description carefully to verify which one provides the tools you need for the task.
+       Each MCP has a configuration that must be accurately implemented in the agent configuration via MCPStdio().
+       Always suggest only the minimum subset of tools from the MCP server URL that are necessary for the solving the task at hand.
+       If the agent is required to generate any intermediate files, you may ask it to save them in a path relative to the current working directory (do not give absolute paths).
+       You must never import or assign `search_mcp_servers` to the tools list of the generated agent in `agent_code`.
 
 #### Structured Output (output_type):
 - Define Pydantic v2 models to structure the agent's final output
 - Implement the `output_type` argument correctly to obtain this structured response
+
+#### Agent Trace (agent_trace): Conditional on the whether the agent code requested is run via CLI or A2AServing
+Important: Saving agent_trace is ONLY required when running the agent via CLI with `agent.run()`. You MUST NEVER save the agent trace when running the agent via A2AServing.
+If the code corresponds to running the agent via CLI, use the following instructions to save the agent trace:
+- Include the agent trace being saved into a JSON file named `agent_eval_trace.json` immediately after agent.run()
+- Saving of the agent trace in the code should be done to the `script_dir / "agent_eval_trace.json"` directory as shown in the example code
+- You would accomplish this by including the lines agent_trace.model_dump_json(indent=2) as shown in the example code
+- Never try to print, log or access any other properties of the agent trace object. agent_trace.response or agent_trace.output are invalid
+- Only agent_trace.model_dump_json(indent=2) and agent_trace.final_output are valid
+- Do not print or save anything after saving the agent trace
 
 ### Code Organization
 - Create well-documented, modular code with appropriate comments
@@ -293,24 +410,19 @@ using Mozilla's any-agent library. The implementation should:
   `.env` file:
     - Use Python `dotenv` library to load the environment variables and access them using
       `os.getenv()`
-
 ### Agent code template
 
 - Rely on the following template to write the agent code:
 
-```
-{AGENT_CODE_TEMPLATE}
-```
 """  # noqa: E501
 
 # Define the template with Jinja2 syntax
 INSTRUCTIONS_TEMPLATE = """
-You are an expert software developer with a deep understanding of Mozilla AI's any-agent Python
-library.
+You are an expert software developer with a deep understanding of Mozilla AI's any-agent Python library.
 
 Any-agent library enables you to:
 - Build agent systems with a unified API regardless of the underlying framework
-- Switch between different agent frameworks (e.g., OpenAI, LangChain, etc) without rewriting code
+- Switch between different agent frameworks (like OpenAI, LangChain, smolagents) without rewriting code
 - Create both single-agent and multi-agent systems with consistent patterns
 - Leverage built-in tools like web search and webpage visiting as well as MCP servers
 - Implement comprehensive tracing and evaluation capabilities
@@ -319,22 +431,33 @@ Any-agent library enables you to:
 
 {{ code_generation_instructions }}
 
+{{ agent_code_template }}
+
+{{ agent_code_template_run_option}}
+
 As input to the `AgentConfig`, you are required to provide the parameters `model_id`,
 `instructions`, `tools`, and `output_type`.
 You also need to specify the correct imports, which have to be consistent with the tools used by the
 agent:
-
 {{ code_example_with_comments }}
+
+{{ code_example_run_option }}
 
 ** Deliverables Instructions**
 
 {{ deliverables_instructions }}
 """  # noqa: E501
 
-# Render the template
-template = Template(INSTRUCTIONS_TEMPLATE)
-INSTRUCTIONS = template.render(
-    code_generation_instructions=CODE_GENERATION_INSTRUCTIONS,
-    code_example_with_comments=CODE_EXAMPLE_WITH_COMMENTS,
-    deliverables_instructions=DELIVERABLES_INSTRUCTIONS,
-)
+
+def load_system_instructions(for_cli_agent: bool = True):
+    template = Template(INSTRUCTIONS_TEMPLATE)
+    return template.render(
+        code_generation_instructions=CODE_GENERATION_INSTRUCTIONS,
+        agent_code_template=AGENT_CODE_TEMPLATE,
+        agent_code_template_run_option=AGENT_CODE_TEMPLATE_RUN_VIA_CLI
+        if for_cli_agent
+        else AGENT_CODE_TEMPLATE_RUN_VIA_A2A,
+        code_example_with_comments=CODE_EXAMPLE_WITH_COMMENTS,
+        code_example_run_option=CODE_EXAMPLE_RUN_VIA_CLI if for_cli_agent else CODE_EXAMPLE_RUN_VIA_A2A,
+        deliverables_instructions=DELIVERABLES_INSTRUCTIONS,
+    )
