+++ src/agent/utils/artifact_validation.py
@@ -1,0 +1,17 @@
+import autoflake
+
+from agent.utils.logging import logger
+
+
+def clean_python_code_with_autoflake(code: str) -> str:
+    """Clean Python code using autoflake to remove unused imports (F401) and variables (F841)."""
+    try:
+        cleaned_code = autoflake.fix_code(
+            code,
+            remove_all_unused_imports=True,
+            remove_unused_variables=True,
+        )
+        return cleaned_code
+    except Exception as e:
+        logger.error(f"Error while running autoflake to clean Python code: {e}")
+        raise
