+++ src/agent/utils/io_utils.py
@@ -1,0 +1,17 @@
+import uuid
+from datetime import datetime
+from pathlib import Path
+
+
+def setup_output_directory(output_dir: Path | None = None) -> Path:
+    if output_dir is None:
+        output_dir = Path.cwd()
+        uid = datetime.now().strftime("%Y-%m-%d_%H:%M:%S") + "_" + str(uuid.uuid4())[:8]
+        # store in a unique dir under generated_workflows by default
+        output_dir = output_dir / "generated_workflows" / uid
+    else:
+        # guarantee output_dir is PosixPath
+        output_dir = Path(output_dir)
+
+    output_dir.mkdir(parents=True, exist_ok=True)
+    return output_dir
