+++ src/agent_factory/generation.py
@@ -1,36 +1,21 @@
 import json
-import uuid
-from datetime import datetime
 from pathlib import Path
 
-import autoflake
 import dotenv
 import fire
 from any_agent import AgentConfig, AgentFramework, AgentRunError, AnyAgent
 from any_agent.tools import search_tavily, visit_webpage
 from any_agent.tracing.agent_trace import AgentTrace
-from pydantic import BaseModel, Field
 
-from agent_factory.instructions import AGENT_CODE_TEMPLATE, INSTRUCTIONS
-from agent_factory.logging import logger
+from agent.factory_tools import read_file, search_mcp_servers
+from agent.instructions import AGENT_CODE_TEMPLATE, AGENT_CODE_TEMPLATE_RUN_VIA_CLI, load_system_instructions
+from agent.schemas import AgentFactoryOutputs
+from agent.utils.artifact_validation import clean_python_code_with_autoflake
+from agent.utils.io_utils import setup_output_directory
+from agent.utils.logging import logger
 from agent_factory.prompt import UserPrompt
-from agent_factory.tools import read_file, search_mcp_servers
 
 dotenv.load_dotenv()
-
-
-class AgentFactoryOutputs(BaseModel):
-    agent_instructions: str = Field(..., description="The instructions passed to the generated agent.")
-    tools: str = Field(..., description="The python code that defines the tools to be used by the generated agent.")
-    imports: str = Field(..., description="The python code snippet used to import the required tools.")
-    structured_outputs: str = Field(..., description="The Pydantic v2 models used to structure the agent's output.")
-    cli_args: str = Field(..., description="The arguments to be provided to the agent from the command line.")
-    agent_description: str = Field(..., description="The description of the agent and what it does.")
-    prompt_template: str = Field(
-        ..., description="A prompt template that, completed with cli_args, defines the agent's input prompt."
-    )
-    run_instructions: str = Field(..., description="The run instructions in Markdown format")
-    dependencies: str = Field(..., description="The dependencies line by line in Markdown format")
 
 
 def create_agent():
@@ -39,7 +24,7 @@ def create_agent():
         framework,
         AgentConfig(
             model_id="o3",
-            instructions=INSTRUCTIONS,
+            instructions=load_system_instructions(for_cli_agent=True),
             tools=[visit_webpage, search_tavily, search_mcp_servers, read_file],
             output_type=AgentFactoryOutputs,
             model_args={"tool_choice": "required"},  # Ensure tool choice is required
@@ -68,34 +53,6 @@ def build_run_instructions(user_prompt) -> str:
     else:
         user_prompt_instance = UserPrompt._instance
         return user_prompt_instance.amend_prompt(user_prompt)
-
-
-def clean_python_code_with_autoflake(code: str) -> str:
-    """Clean Python code using autoflake to remove unused imports (F401) and variables (F841)."""
-    try:
-        cleaned_code = autoflake.fix_code(
-            code,
-            remove_all_unused_imports=True,
-            remove_unused_variables=True,
-        )
-        return cleaned_code
-    except Exception as e:
-        logger.error(f"Error while running autoflake to clean Python code: {e}")
-        raise
-
-
-def setup_output_directory(output_dir: Path | None = None) -> Path:
-    if output_dir is None:
-        output_dir = Path.cwd()
-        uid = datetime.now().strftime("%Y-%m-%d_%H:%M:%S") + "_" + str(uuid.uuid4())[:8]
-        # store in a unique dir under generated_workflows by default
-        output_dir = output_dir / "generated_workflows" / uid
-    else:
-        # guarantee output_dir is PosixPath
-        output_dir = Path(output_dir)
-
-    output_dir.mkdir(parents=True, exist_ok=True)
-    return output_dir
 
 
 def run_agent(agent: AnyAgent, user_prompt: str, max_turns: int = 30) -> AgentTrace:
@@ -128,7 +85,11 @@ def save_agent_outputs(agent_trace: AgentTrace, output_dir: Path) -> None:
         agent_path = output_dir / "agent.py"
         instructions_path = output_dir / "INSTRUCTIONS.md"
         requirements_path = output_dir / "requirements.txt"
-        agent_code = AGENT_CODE_TEMPLATE.format(**agent_trace.final_output.model_dump())
+        agent_code = (
+            f"{AGENT_CODE_TEMPLATE.format(**agent_trace.final_output.model_dump())} \n"
+            f"{AGENT_CODE_TEMPLATE_RUN_VIA_CLI.format(**agent_trace.final_output.model_dump())}"
+        )
+
         cleaned_agent_code = clean_python_code_with_autoflake(agent_code)
 
         with agent_path.open("w", encoding="utf-8") as f:
