+++ tests/generation/unit/test_generation_utils.py
@@ -3,18 +3,18 @@ from unittest.mock import MagicMock, patch
 import pytest
 from any_agent import AgentRunError, AgentTrace
 
+from agent.utils.io_utils import setup_output_directory
 from agent_factory.generation import (
     run_agent,
-    setup_output_directory,
 )
 
 
 def test_setup_output_directory_with_none_creates_unique_directory(tmp_path):
     """Ensures that passing None as output_dir creates a unique directory."""
-    with patch("agent_factory.generation.Path.cwd", return_value=tmp_path):
-        with patch("agent_factory.generation.datetime") as mock_datetime:
+    with patch("agent.utils.io_utils.Path.cwd", return_value=tmp_path):
+        with patch("agent.utils.io_utils.datetime") as mock_datetime:
             mock_datetime.now.return_value.strftime.return_value = "2025-01-01_12:00:00"
-            with patch("agent_factory.generation.uuid.uuid4") as mock_uuid:
+            with patch("agent.utils.io_utils.uuid.uuid4") as mock_uuid:
                 mock_uuid.return_value.hex = "1234567890abcdef"  # pragma: allowlist secret
 
                 result = setup_output_directory()
